+++ CHANGELOG.md
@@ -4,11 +4,12 @@
 
 ### Features
 
-Updates the `Alert` component design and adds new optional features: headline, close button, and learn-more link
+Updates the `Alert` component design and adds new optional features: headline, close button, and link
 
 ### Breaking changes
 
 - the `linkColor` functionality in the Alert content is deprecated
+
 ## v2.0.0-beta.24
 
 ### Bug Fixes